all: change import path to go4.org/grpc
I want to continue the grpc-on-x/net/http2 experiment but also change
the code generator to remove some types. So, give it a distinct import
path so the clients wanting to use the experimental version can
compile in the same workspace as clients using the official
google.golang.org/grpc.